Add on-demand topic page generation and extract renderTopicHtml utility
When a /topic/:term URL is visited and no static page exists yet, the server
now generates and saves it on the fly so Google can crawl any organically
shared link, not just pre-seeded ones from the startup script.

- Extract renderTopicHtml() into utils/renderTopicHtml.js (shared module)
- generate_static_topics.js imports from shared util (was self-contained)
- index.js: /topic/:term route generates+saves page on first visit with
  in-memory deduplication guard to prevent concurrent duplicate writes
- /sitemap.xml now scans dist/topic/ directory so on-demand pages appear
  automatically in the sitemap on next crawl
